minor folder restructuring
