+++ src/components/indicator/timer/Timer.js
