+++ src/components/indicator/timer/index.js
